fix: renamed button file and changed import statements
# Conflicts:
#	frontend/src/containers/pages/ErrorPages/InvalidRolePage.jsx
