+++ frontend/src/components/TryAccessScenarioButton.jsx
